Create an abstract class
The bot code was modified so that the presentation of information to the user (displaying cards with user contacts, notes, a page with information about available commands) could be easily changed. For this purpose, an abstract base class for user views and specific implementations that inherit the base class and implement the console interface are provided.